chore: fix Rust 1.92 clippy warnings
- Fix new clippy lints across snapshot harnesses and tests (identity_op, needless_return, doc_lazy_continuation, default_constructed_unit_structs)
- Use Error::other/is_multiple_of helpers and add is_empty for the wasm OPFS backend
- Minor wasm audio bridge cleanups

This keeps CI (clippy -D warnings) green on Rust 1.92 without running repo-wide rustfmt.

+++ crates/platform/src/audio/worklet_bridge.rs
@@ -487,7 +487,7 @@ mod wasm {
 
             let cc = self.channel_count as usize;
             let first_samples = first_frames as usize * cc;
-            let write_sample_pos = write_pos as u32 * self.channel_count;
+            let write_sample_pos = write_pos * self.channel_count;
             self.samples
                 .subarray(write_sample_pos, write_sample_pos + first_samples as u32)
                 .copy_from(&samples[..first_samples]);
